Made some changes
- changed the data type of every loop control variable in a for loop
- made other changes on input_info.cpp and nth_power_number_checker.cpp

+++ cryptography/text_to_morse_code.cpp
@@ -19,7 +19,7 @@ int main() {
 
     cout << endl << "Morse Code: " << endl;
 
-    for (int counter = 0; counter < length; counter++) {
+    for (long long counter = 0; counter < length; counter++) {
 
         char character = text[counter];
