+++ number_tools/armstrong_number_checker.cpp
@@ -7,7 +7,7 @@ using namespace std;
 
 int main() {
 
-    int num = 0, sum = 0, length = 0, digit;
+    long long num = 0, sum = 0, length = 0, digit;
     string str_num;
 
     cout << "Number: ";
@@ -16,7 +16,7 @@ int main() {
     str_num = to_string(num);
     length = str_num.length();
 
-    for (int counter = 0; counter < length; counter++) {
+    for (long long counter = 0; counter < length; counter++) {
         digit = str_num[counter] - '0';
         sum += pow(digit, length);
     }
