+++ number_tools/factorial_finder.cpp
@@ -5,12 +5,12 @@ using namespace std;
 
 int main() {
 
-    int num, factorial = 1;
+    long long num, factorial = 1;
 
     cout << "Number: ";
     cin >> num;
 
-    for (int counter = 1; counter <= num; counter++) {
+    for (long long counter = 1; counter <= num; counter++) {
         factorial *= counter;
     }
