+++ number_tools/prime_number_checker.cpp
@@ -5,12 +5,12 @@ using namespace std;
 
 int main() {
 
-    int num, largest_factor = 1;
+    long long num, largest_factor = 1;
 
     cout << "Number: ";
     cin >> num;
 
-    for (int counter = 2; counter < num; counter++)
+    for (long long counter = 2; counter < num; counter++)
         if (num % counter == 0)
             largest_factor = counter;
