+++ text_tools/reverse_text_alignment.cpp
@@ -15,7 +15,7 @@ int main() {
 
     cout << endl << "Reversed String: " << endl;
 
-    for (int counter = length - 1; counter >= 0; counter--)
+    for (long long counter = length - 1; counter >= 0; counter--)
         cout << text[counter];
 
     return 0;
