+++ text_tools/text_pyramid.cpp
@@ -19,8 +19,8 @@ int main() {
 
     if (alignment == "L") {
 
-        for (int counter_1 = 1; counter_1 <= num; counter_1++) {
-            for (int counter_2 = 1; counter_2 <= counter_1; counter_2++)
+        for (long long counter_1 = 1; counter_1 <= num; counter_1++) {
+            for (long long counter_2 = 1; counter_2 <= counter_1; counter_2++)
                 cout << triangle_piece;
             cout << endl;
         }
@@ -36,13 +36,13 @@ int main() {
             num--;
         }
 
-        for (int counter_1 = 1; counter_1 <= num / 2 + 1; counter_1++) {
+        for (long long counter_1 = 1; counter_1 <= num / 2 + 1; counter_1++) {
 
-            for (int counter_2 = 1; counter_2 <= spaces/2; counter_2++)
+            for (long long counter_2 = 1; counter_2 <= spaces/2; counter_2++)
                 cout << " ";
-            for (int counter_2 = 1; counter_2 <= asterisks; counter_2++)
+            for (long long counter_2 = 1; counter_2 <= asterisks; counter_2++)
                 cout << triangle_piece;
-            for (int counter_2 = 1; counter_2 <= spaces/2; counter_2++)
+            for (long long counter_2 = 1; counter_2 <= spaces/2; counter_2++)
                 cout << " ";
 
             asterisks += 2;
@@ -57,11 +57,11 @@ int main() {
 
 		spaces = num - 1;
 
-        for (int counter_1 = 1; counter_1 <= num; counter_1++) {
+        for (long long counter_1 = 1; counter_1 <= num; counter_1++) {
 
-            for (int counter_2 = 1; counter_2 <= spaces; counter_2++)
+            for (long long counter_2 = 1; counter_2 <= spaces; counter_2++)
                 cout << " ";
-            for (int counter_2 = 1; counter_2 <= asterisks; counter_2++)
+            for (long long counter_2 = 1; counter_2 <= asterisks; counter_2++)
                 cout << triangle_piece;
 
 			asterisks++;
